Extract CLSInstance into a separate file. Add unit tests
